Add AddItem function (#1)
* Add azure function to add a new shopping list item to Cosmos DB

Update unit model for shopping list bot to enable unitName as optional and add function project for an Azure Function to add a new shopping list item to a Cosmos DB using the Cosmos DB connector

* Add request to AddItemFunction from bot

Update bot to call AddItemFunction component if a item should be added. Introduce a functions service for the API calls.