Fix adaptive cards callback
All items adaptive card save did not work because the id of the items was not set

Remove item from Update Item Function because it would only return the item we already know in most cases as such it does not help the bot.

TODO QueryItemIdDialog could be refactored to not just return the id of the item found but the entire item. The item could then be used in the error message or success message of the mark and unmark dialogs.